requirements.txt
langgraph==0.2.28
langchain==0.3.0
langchain-openai==0.2.0
pydantic==2.9.2
python-dotenv==1.0.1
